+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add support of Markdown [#679](https://github.com/nbbrd/spreadsheet4j/issues/679)
+- Add support of Excel Binary Workbook (*.xlsb) [#9](https://github.com/nbbrd/spreadsheet4j/issues/9)
 
 ### Fixed